+++ Blurred-ios/Channels/Video/Overlay/OverlayViewDelegate.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 protocol ChannelVideoOverlayViewDelegate: AnyObject {
-    func didTapChannel(_ sender:ChannelVideoOverlayView)
-    func didTapComments(_ sender:ChannelVideoOverlayView)
+    func didTapChannel(_ view:ChannelVideoOverlayView, videousername: String)
+    func didTapComments(_ view:ChannelVideoOverlayView, videoid: Int)
 }